requirements.txt
Flask
Flask-MySQLdb
PyYAML
Werkzeug
mysql-connector-python
beautifulsoup4
requests
